TempmailBot/requirements.txt
aiogram 
aio-pika
aiosmtpd 
dkimpy 
tenacity
django
dj-database-url
psycopg2-binary 
django-unfold 
django-import-export